creating students in database

+++ internal/storage/storage.go
@@ -1,5 +1,5 @@
 package storage
 
 type Storage interface {
-	CreateStudent(name string, email string, age int) (uint32, error)
+	CreateStudent(name string, email string, age int) (int64, error)
 }
